requirements.txt
streamlit==1.43.1
streamlit-webrtc
torch
torchvision
Pillow
opencv-python-headless
requests
numpy
